+++ src/components/Main/Main.module.css
@@ -3,3 +3,10 @@
   grid-template-columns: 1fr 3fr;
   grid-gap: 5px;
 }
+
+@media (max-width: 1024px) {
+  .main {
+    grid-template-columns: 1fr;
+    grid-gap: 0;
+  }
+}
