+++ src/components/Main/Nav/Friends/Friend/Friend.module.css
@@ -1,6 +1,6 @@
 .person {
   background: #eeeeee;
-  width: 17rem;
+  width: 60%;
   padding: 1rem;
   margin: 1rem auto;
   transition: all 0.3s linear;
