+++ src/components/Main/Nav/Friends/Friends.module.css
@@ -8,3 +8,9 @@
   font-size: 24px;
   text-align: center;
 }
+
+@media (max-width: 1024px) {
+  .friends {
+    display: none;
+  }
+}
