+++ src/components/Main/Nav/Nav.jsx
@@ -1,24 +1,15 @@
 import React from "react";
-import { NavLink } from "react-router-dom";
 import Friends from "./Friends/Friends";
+import MobileNavigation from "./MobileNavigation/MobileNavigation";
 import classes from "./Nav.module.css";
+import Navigation from "./Navigation/Navigation";
 
 const Nav = ({ friends, navLinks }) => {
-  const SetClassName = (navData) => {
-    return navData.isActive ? classes.active : classes.link;
-  };
   return (
     <div className={classes.sideBar}>
-      <nav className={classes.nav}>
-        {navLinks.map((link) => (
-          <div className={classes.item} key={link.id}>
-            <NavLink to={link.to} className={SetClassName} alt={link.alt}>
-              {link.name}
-            </NavLink>
-          </div>
-        ))}
-      </nav>
-      <Friends friends={friends} />;
+      <Navigation navLinks={navLinks} />
+      <MobileNavigation navLinks={navLinks} />
+      <Friends friends={friends} />
     </div>
   );
 };
