+++ src/components/Main/pages/Profile/Posts/Posts.module.css
@@ -1,3 +1,9 @@
 .posts {
   margin: 10px;
 }
+
+@media (max-width: 1024px) {
+  .posts {
+    margin: 25px;
+  }
+}
